Merge Naomi's relevant changes from summer internship into develop branch (#6)
* case_study_flux line 56 input changed

filename --> filenames because of error with input name

* case_study_flux_tests calling Field update

update to as is in develop to avoid type error

* Update filenames (to develop names)

* Fix indentation error

* Indentation error part 2

* Add files via upload

* Add files via upload

* Create trying

* Add files via upload

* Delete case_settings.csv

* Delete trying

* Add files via upload

* Add files via upload

* Create vd

* Add files via upload

* Delete vd

* Update PS_10_ca_case.csv

* Update extcyl_PS-10.csv

* Add line to fix key error

* Fix mirror_len and mirror_ht

* Add diameter and change vals

* Add line to fix key error

* Add line to fix key error

* Hour_id fixes

* Add extra input to fix reading weather type error for SP

* Reduce number of lines of code

in read_weather statement

* Add option to create plot overlapping measurement and aim points

Option is passed into main.py and is then used as input into functions in solve_aim_model, process_aimpoint_outputs, plotting, and flux_model

* Add the types of files made in the building process to .gitignore

So I don't have to stash and discard so many changes. Added png, csv, and pyc.

* Put copylot in code folder

So that user doesn't have to download it and so it doesn't get deleted when making changes between local and remote

* Revert "Add the types of files made in the building process to .gitignore"

This reverts commit 731296ec7e1d844edbdf8315a061f8f22edf2ba3.

* Changes to make measurement graph plotting simpler coding

Realized my first way was overly complicated and that this is a better way to make the plot. Made it so it always makes this graph, no longer pass input in to make it.

* delete unused plot_meas input

* Put solarpilot.dII in folder

* Case settings for running flat 50 aimpoint cases

Same as settings in develop except for different hour and field DNI, as well as the inclusion of diameter of 10.38. Need to change aim_cols and aim_rows as desired.

* Fix plot measurement points so works for flat plate too

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Delete building file

* Add post-refocusing functionality

Search for and implement feasible heliostat-aimpoint pairings for defocused heliostats after optimization. Primarily involves changes/additions to process_aimpt and optimize_aimpt files.

Also option to pass in easily pass in big summary file to record the results of cases run in a row.

* Refocus fix

Changed variables to pass through from optimize_aimpt to process_aimpt

* Added comments and slight changes to refocus code

* Change solver to cbc

* Set beam as DNI for INTL files (Chile file)

* Rename Spain file to IWEC type, not INTL

* Add additional strategies: CenterOut algorithmic, weighted center, flux size weighted

* Fix so that heliostat refocuses only at the first aimpoint for which it can

add break in loop

* Fix measurement plotting to work for cylindrical and flat plate

* Fix plotting error with flat dimensions

* Calculate initial feasible objective value and time taken; record in summary and solutions file

* Updated FluxSize -- break into more flux size sections

got 84884 obj val for 10 tries with this version of code

* Updates to opt_heuristic

This version yielded a 2% increase for FluxSize. Main change is if there are defocused heliostats in the section's obtained solution to reassign the h-a pairings randomly.

* Changes to opt_heuristic

Deleted some strategies I wrote and edited the strategy that I'm keeping (FluxStoreSize)

* In AcceptReject, check for switching after

option for both random or by sorting by flux size and going in order

* Add post-initial solution switching aimpoints

* Store results from initial heuristic in summary file

* case_study_flux line 56 input changed

filename --> filenames because of error with input name

* Create solarpilot

* case_study_flux_tests calling Field update

update to as is in develop to avoid type error

* Put copylot in code folder

So that user doesn't have to download it and so it doesn't get deleted when making changes between local and remote

* Delete unnecessary changes to inputs and added files

* Minor clarifications and edits

* Store results from initial heuristic in summary file

* Return initial values and print in process file

* Small changes to make easier for general use

* Delete a couple building files

* Update number of aimpoints to default value

* Delete building file

* Revert "Delete a couple building files"

This reverts commit 47a3dc7c0fcd56de0362d73b49360e70d65abf8f.

* Delete unused import line in sp_module

* Add docstrings to functions, reformat imports, delete copylot.py and solarpilot.dII